added object package and sprites

+++ src/entity/Entity.java
@@ -4,7 +4,7 @@ import java.awt.*;
 import java.awt.image.BufferedImage;
 
 public class Entity {
-    public double worldX,worldY;
+    public int worldX,worldY;
     public int speed;
 
     public BufferedImage up1,up2,down1,down2,left1,right1,left2,right2;
